requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
idna==3.6
numpy==1.26.4
pillow==10.2.0
pycparser==2.21
pynput==1.7.6
pywin32==225
requests==2.31.0
scipy==1.12.0
secure-smtplib==0.1.1
six==1.16.0
sounddevice==0.4.6
urllib3==2.2.0
